+++ modules/siddhi-extensions/input-transports/jms-input-transport/src/main/java/org/wso2/siddhi/extension/input/transport/jms/JMSMessageProcessor.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2016 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *  Copyright (c) 2017 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  *  WSO2 Inc. licenses this file to you under the Apache License,
  *  Version 2.0 (the "License"); you may not use this file except
@@ -16,9 +16,8 @@
  *  under the License.
  *
  */
-package org.wso2.siddhi.extension.input.transport.jms.util;
+package org.wso2.siddhi.extension.input.transport.jms;
 
-import org.apache.log4j.Logger;
 import org.wso2.carbon.messaging.CarbonCallback;
 import org.wso2.carbon.messaging.CarbonMessage;
 import org.wso2.carbon.messaging.CarbonMessageProcessor;
